rename: DashboardLayout -> DashBoardLayout 네이밍 변경
